model_migration_eval/requirements.txt
# Azure OpenAI SDK
openai>=1.40.0
azure-identity>=1.15.0

# Azure Blob Storage (user data persistence)
azure-storage-blob>=12.19.0

# Microsoft Foundry Control Plane Evaluation (optional)
azure-ai-projects>=2.0.0b2

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-compress>=1.15

# Data Processing
numpy>=1.24.0
pyyaml>=6.0.1

# Evaluation & Metrics
scikit-learn>=1.3.0

# Tokenizer (prompt health analysis)
tiktoken>=0.7.0

# Async Support
httpx>=0.26.0

# WebSocket support for Realtime API (speech-to-speech)
websockets>=12.0

# Caching
diskcache>=5.6.3

# Environment & Config
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
